Check for nil when slicing array
Sometimes the indices passed to slice don't match the range of the
buffer. This can happen during complex chains of edits, such as an undo
sequence. Ideally, we would ensure that the indices passed to slice are
always valid, but a simpler solution that at least prevents errors
caused from using a nil value is to just ensure that the last row is
non-nil.